requirements.txt
streamlit==1.25.0
google-generativeai
numpy==1.24.3
matplotlib==3.8.0
pandas==1.5.3
scikit-learn==1.3.0
PyPDF2==3.0.1